% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesCreate.R
\name{BcolsFromBcolllist}
\alias{BcolsFromBcolllist}
\title{TODO}
\usage{
BcolsFromBcolllist(Bcollist)
}
\arguments{
\item{Bcollist}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")
}
