% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesCreate.R
\name{BsFromB}
\alias{BsFromB}
\title{TODO}
\usage{
BsFromB(B, s = NULL, r = NULL, permlist = NULL, oneonly = TRUE)
}
\arguments{
\item{B}{TODO}

\item{s}{TODO}

\item{r}{TODO}

\item{permlist}{TODO}

\item{oneonly}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("tODO")
}
