% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.FFsym2.R
\name{contr.FFsym2}
\alias{contr.FFsym2}
\title{Full-factorial-based Helmert type contrasts for s^el levels}
\usage{
contr.FFsym2(n, s, contrasts = TRUE, slowfirst = TRUE)
}
\arguments{
\item{n}{integer or vector; either an integer number of levels of the factor for
which contrasts are created, which must be a a power of \code{s}; or a factor
whose number of levels is a power of \code{s}; or a vector of levels whose
number of elements is a power of \code{s}.}

\item{s}{positive integer, at least 2}

\item{contrasts}{logical; must be TRUE}

\item{slowfirst}{logical; default TRUE}
}
\value{
\code{contr.FFsym2} yields a matrix of real-valued contrasts.
That matrix can be used in function \code{model.matrix}
or in any statistical modeling functions.
}
\description{
Full-factorial-based Helmert type contrasts for s^el levels
}
\details{
The function implements real-valued contrasts that can be used
instead of the complex-valued contrasts from
Tian and Xu (2022), as implemented in function
\code{\link{contr.TianXu}}. Its main use is the calculation of the
stratification pattern (also called space-filling pattern).
}
\examples{
## the same n can yield different contrasts for different s
contr.FFsym(16, 2)
round(contr.FFsym(16, 4), 4)
round(contr.FFsym(16, 16), 4)

}
\references{
Tian and Xu (2022)
}
