% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShiTang.R
\name{create_ABC}
\alias{create_ABC}
\title{Create ABC object}
\usage{
create_ABC(k, m = NULL, constr = "ShiTang_alphabeta")
}
\arguments{
\item{k}{the log base 2 of the run size (n) of the SOA.  \code{k} must be greater or equal to 4}

\item{m}{TODO}

\item{constr}{type of construction.  Must be one of \code{"ShiTang_alphabeta", "ShiTang_alpha"}}
}
\value{
A list of components
\describe{
\item{A}{}
\item{B}{}
\item{C}{}
\item{Yates.columns}{A list with components \code{A, B, C}}
}
}
\description{
Create ABC object
}
\examples{
M <- create_ABC(4)
}
