% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShiTang.R
\name{create_DfromABC}
\alias{create_DfromABC}
\title{Create object D from an ABC object}
\usage{
create_DfromABC(listABC, permlist = NULL, random = FALSE, ...)
}
\arguments{
\item{listABC}{an object with elements A, B, C and Yates.columns with elements A,B,C}

\item{permlist}{a list of length m with three elements each}

\item{random}{is the construction randomized?}

\item{...}{Not used}
}
\value{
a matrix
}
\description{
Create object D from an ABC object
}
\keyword{internal}
