% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesCreate.R
\name{createAB}
\alias{createAB}
\title{TODO}
\usage{
createAB(s, k = 3, m = NULL)
}
\arguments{
\item{s}{TODO}

\item{k}{TODO}

\item{m}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")

}
\keyword{internal}
