% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createABCcols.R
\name{createABCcols}
\alias{createABCcols}
\title{Another utility for array creation (Shi and Tang)}
\usage{
createABCcols(k)
}
\arguments{
\item{k}{integer; determines the run size: the resulting array will have s^k runs}
}
\value{
\code{createABCcols} returns a list of Yates matrix column numbers to
be used for A, B, AB and C for the Shi and Tang (2020) family 1 construction
}
\description{
Another utility for array creation (Shi and Tang)
}
\keyword{internal}
