% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShiTang.R
\name{createABcols}
\alias{createABcols}
\title{initialize recursive construction of A,B,C}
\usage{
createABcols(k)
}
\arguments{
\item{k}{the log base 2 of the run size (n) of the SOA.  \code{k >= 4}}
}
\value{
a list containing
\describe{
\item{Acols}{}
\item{Bcols}{}
\item{Ccols}{}
}
}
\description{
Used in the Shi and Tang Family 1 construction
}
\keyword{internal}
