% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiaoXuMDLE.R
\name{createF}
\alias{createF}
\title{TODO}
\usage{
createF(Dc, Dp, s, ell, nseq = 2000)
}
\arguments{
\item{Dc}{TOIO}

\item{Dp}{TODO}

\item{s}{TODO}

\item{ell}{TODO}

\item{nseq}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")

}
\keyword{internal}
