% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesCreate.R
\name{createSaturated}
\alias{createSaturated}
\title{TODO}
\usage{
createSaturated(s, k = 2)
}
\arguments{
\item{s}{TODO}

\item{k}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")
}
