% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDLEs.R
\name{DcFromDp}
\alias{DcFromDp}
\title{Create the expansions}
\usage{
DcFromDp(
  Dp,
  s,
  ell,
  permlist = rep(list(rep(list(rep(0:(ell - 1), each = nrow(Dp)/(s * ell))), s)),
    ncol(Dp))
)
}
\arguments{
\item{Dp}{an OA (ideally maximin optimized GMA OA)}

\item{s}{the number of levels of each column in Dp}

\item{ell}{\code{ell*s} is the target number of levels of the outgoing Dc}

\item{permlist}{a list of m lists of length s permutations of the elements
of the replacement vector}
}
\value{
a matrix of the same size as \code{Dp}
}
\description{
this function is used in each step of the Weng optimization and for
the initialization of the XiaoXu optimization
}
\examples{
DcFromDp(matrix(c(0,0,1,1,0,1,0,1), nrow = 4), 2, 2)
}
