% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSP.R
\name{fastSP}
\alias{fastSP}
\alias{fastSP.K}
\title{Functions for fast calculation of stratification pattern
according to Tian and Xu 2023}
\usage{
fastSP(D, s, maxwt = NULL, K = NULL, y0 = 0.1, tol = 1e-05)

fastSP.K(D, s, maxwt = NULL, y0 = 0.1, K = NULL, tol = 1e-05)
}
\arguments{
\item{D}{design with number of levels a power of \code{s}}

\item{s}{prime or prime power on which \code{D} is based}

\item{maxwt}{maximum weight for which the pattern is to be calculated}

\item{K}{integer number of summands in Theorems 3 and 4 of Tian and Xu (2023+) for
approximate calculation of the first few elements of the stratification pattern.
Larger \code{K} provide better accuracy. Default (\code{NULL}) is the maximum of
\code{maxwt} (if specified) and a default based on \code{y0} according to a formula
by Tian and Xu (2023+) (yields smaller \code{K} for smaller \code{y0}).\cr
\code{K} should not divide \code{m*el}, according to Tian and Xu (2023+).
??? in the example, 9 works well, though it divides 9*3
sent a question to Hongquan regarding the rationale behind this.\cr
In function \code{fastSP}, for obtaining all summands of Theorem 3 of
Tian and Xu (2023+), \code{K="max"} can be specified.}

\item{y0}{small number that drives accuracy. Tian and Xu (2023+) proposed
\code{y0}=0.1 for the approximation of \code{fastSP.K}, and generally choices
between 0.001 and 0.1. As \code{y0}=0.1 appears to improve
performance also for \code{fastSP}, it is introduced here also for that function.
\code{maxwt} and a formula by Tian and Xu (2023+) that depends on}

\item{tol}{tolerance for checking whether the imaginary part is zero}
}
\value{
\code{fastSP} returns a stratification pattern or the first
\code{maxwt} elements of the stratification pattern (default: all elements).\cr
If \code{K} is less than the maximum length of the stratification pattern
(number of columns of
\code{D} times exponent of \code{s} that yields the number of levels)
the returned values are approximations (more accurate for larger \code{K}).
}
\description{
Functions for fast calculation of stratification pattern
according to Tian and Xu 2023
}
\details{
The functions were modified from the code provided with
Tian and Xu (2023+).

Function \code{fastSP} implements the formula of Theorem 3 of
Tian and Xu (2023+), but with yj replaced with yj*y0^j, as this appears to
yield more accurate results (as evidenced by the fact that it worked for the
20-column SOA for which the original version failed to yield accurate results).\cr
Per default (\code{maxwt=NULL} and \code{K=NULL}),
\code{fastSP} calculates the entire stratification pattern and uses all summands
of Tian and Xu's Equation (4).\cr
It is possible and often advisable to calculate only a smaller number of
entries, for saving resources and also because the later entries are less
interesting and less accurate. If the argument \code{maxwt} is specified,
the default for the argument \code{K} (\code{K=NULL}) is the maximum of
Tian and Xu's (2023+) proposed default for their approximation formula (6)
(depending on \code{y0}) and the requested \code{maxwt}.\cr
For obtaining the original behavior of Tian and Xu's (2023+) implementation
of their Theorem 3 (not desirable for larger situations),
choose \code{y0=1} and \code{K="max"}. Note that \code{y0=1} with
\code{K=NULL} yields an error, because the default formula for \code{K}
does not work for \code{y0=1}.
}
\examples{
## SOA(32,9,8,3) from Shi and Tang (2020)
soa32x9 <- t(matrix(c(7,3,6,2,7,3,6,2,4,0,5,1,4,0,5,1,5,1,4,0,5,1,4,0,6,2,7,3,6,2,7,3,
                  7,7,2,2,5,5,0,0,6,6,3,3,4,4,1,1,5,5,0,0,7,7,2,2,4,4,1,1,6,6,3,3,
                  7,5,6,4,3,1,2,0,4,6,5,7,0,2,1,3,7,5,6,4,3,1,2,0,4,6,5,7,0,2,1,3,
                  7,7,4,4,5,5,6,6,2,2,1,1,0,0,3,3,7,7,4,4,5,5,6,6,2,2,1,1,0,0,3,3,
                  7,5,6,4,5,7,4,6,6,4,7,5,4,6,5,7,3,1,2,0,1,3,0,2,2,0,3,1,0,2,1,3,
                  7,1,0,6,3,5,4,2,4,2,3,5,0,6,7,1,5,3,2,4,1,7,6,0,6,0,1,7,2,4,5,3,
                  7,1,2,4,7,1,2,4,2,4,7,1,2,4,7,1,5,3,0,6,5,3,0,6,0,6,5,3,0,6,5,3,
                  7,3,2,6,5,1,0,4,4,0,1,5,6,2,3,7,3,7,6,2,1,5,4,0,0,4,5,1,2,6,7,3,
                  7,1,4,2,3,5,0,6,2,4,1,7,6,0,5,3,3,5,0,6,7,1,4,2,6,0,5,3,2,4,1,7
), nrow=9, byrow = TRUE))

a <- fastSP(soa32x9, 2); round(a,7)

## el=3 is calculated in the function
a <- fastSP.K(soa32x9, 2, maxwt=5); round(a,7)
## will be more accurate
a <- fastSP.K(soa32x9, 2, maxwt=5, y=0.01); round(a,7)
a <- fastSP.K(soa32x9, 2, maxwt=5, y=0.01, K=9); round(a,7)

# example code

}
\references{
For full detail, see \code{\link{SOAs-package}}.

Tian, Y. and Xu, H. (2023+)
}
