% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nchoosek}
\alias{nchoosek}
\alias{levels.no}
\title{Utility functions from DoE.base}
\usage{
nchoosek(n, k)

\method{levels}{no}(xx)
}
\arguments{
\item{n}{number to select from}

\item{k}{number to be selected without replacement}

\item{xx}{matrix or data.frame}
}
\value{
\code{nchoosek} returns a \code{k} times \code{choose(n,k)} matrix
whose columns hold the possible selections in lexicographic order

\code{levels.no} returns a vector of numbers of levels for the columns of \code{xx}
}
\description{
Utility functions from DoE.base
}
\keyword{internal}
