% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesGF.R
\name{gf_matmult}
\alias{gf_matmult}
\title{Multiple matrices over Galois fields}
\usage{
gf_matmult(M1, M2, gf, checks = TRUE)
}
\arguments{
\item{M1}{matrix 1 with elements in \code{0:(q-1)}}

\item{M2}{matrix 2 with elements in \code{0:(q-1)}}

\item{gf}{Galois field object with characteristic \code{q}}

\item{checks}{should checks be performed}
}
\value{
an integer matrix
}
\description{
Multiple matrices over Galois fields
}
