% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesGF.R
\name{gf_prod}
\alias{gf_prod}
\title{Product of two elements of a Galois field}
\usage{
gf_prod(x, y, gf)
}
\arguments{
\item{x}{the first element in \code{0:(q-1)}}

\item{y}{the second element in \code{0:(q-1)}}

\item{gf}{the Galois field of characteristic \code{q}}
}
\value{
the product
}
\description{
Product of two elements of a Galois field
}
\keyword{internal}
