% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesGF.R
\name{gf_sum}
\alias{gf_sum}
\title{Sum of a two elements of a Galois field}
\usage{
gf_sum(x, y, gf)
}
\arguments{
\item{x}{first element in \code{0:(q-1)}}

\item{y}{second element in \code{0:(q-1)}}

\item{gf}{Galois field of characteristic \code{q}}
}
\value{
the sum
}
\description{
Sum of a two elements of a Galois field
}
\keyword{internal}
