% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesGF.R
\name{gf_sum_list}
\alias{gf_sum_list}
\title{Sum over a list of integer vectors in a Galois field}
\usage{
gf_sum_list(ll, gf, checks = TRUE)
}
\arguments{
\item{ll}{a list of integer vectors in \code{0:(q-1)}}

\item{gf}{a Galois field of characteristic \code{q}}

\item{checks}{should input checks be performed}
}
\value{
a integer vector
}
\description{
Sum over a list of integer vectors in a Galois field
}
\keyword{internal}
