% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesGF.R
\name{int2poly}
\alias{int2poly}
\title{Integer to Galois field polynomial}
\usage{
int2poly(x, gf)
}
\arguments{
\item{x}{integer vector with e elements}

\item{gf}{Galois field object}
}
\value{
e x n matrix or length n vector (if x is scalar)
}
\description{
Integer to Galois field polynomial
}
\keyword{internal}
