% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{interleavecols}
\alias{interleavecols}
\title{TODO}
\usage{
interleavecols(A, B)
}
\arguments{
\item{A}{TODO}

\item{B}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")

}
\keyword{internal}
