% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ncol_lb}
\alias{ncol_lb}
\title{TOOD}
\usage{
ncol_lb(s, k, type = "2+")
}
\arguments{
\item{s}{TODO}

\item{k}{TODO}

\item{type}{TODO}
}
\value{
TODO
}
\description{
TOOD
}
\examples{
print("TODO")

}
\keyword{internal}
