% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighbourcalcUniversal.R
\name{NeighbourcalcUniversal}
\alias{NeighbourcalcUniversal}
\title{Function to do level permutations according to Weng's algorithm}
\usage{
NeighbourcalcUniversal(
  funname,
  mperm,
  r,
  ...,
  startperm = NULL,
  allpermlist = NULL,
  neighbordist = 1
)
}
\arguments{
\item{funname}{function that creates the individual (O)SOAs}

\item{mperm}{number of columns of \code{startperm}}

\item{r}{number of rows of \code{startperm}}

\item{...}{arguments for function \code{funname}}

\item{startperm}{matrix with position numbers of level permutations (refers to \code{allpermlist})}

\item{allpermlist}{list of all permutations}

\item{neighbordist}{1 or 2: one- or two-neighbors in Weng's algorithm}
}
\value{
list of arrays and corresponding permutations
}
\description{
Takes a workhorse function and creates random one- or two-neighbors
}
\examples{
print("none")

}
\keyword{internal}
