% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighbourcalcUniversal_random.R
\name{NeighbourcalcUniversal_random}
\alias{NeighbourcalcUniversal_random}
\title{Function to do level permutations according to Weng's algorithm
without using a stored list of all permutations}
\usage{
NeighbourcalcUniversal_random(
  funname,
  mperm,
  r,
  ...,
  curperms = NULL,
  replacement = NULL,
  neighbordist = 1
)
}
\arguments{
\item{funname}{function that creates the individual (O)SOAs}

\item{mperm}{number of columns of \code{startperm}}

\item{r}{number of rows of \code{startperm}}

\item{...}{arguments for function \code{funname}}

\item{neighbordist}{1 or 2: one- or two-neighbors in Weng's algorithm}

\item{startperm}{matrix with position numbers of level permutations (refers to \code{allpermlist})}

\item{allpermlist}{list of all permutations}
}
\value{
list of arrays and corresponding permutations
}
\description{
Takes a workhorse function and creates random one- or two-neighbors
}
\keyword{internal}
