% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiaoXuMDLE.R
\name{optimize}
\alias{optimize}
\title{TODO}
\usage{
optimize(Dc, s, ell, F, nrounds = 50, nsteps = 3000)
}
\arguments{
\item{Dc}{TODO}

\item{s}{TODO}

\item{ell}{TODO}

\item{F}{TODO}

\item{nrounds}{TODO}

\item{nsteps}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("tODO")

}
\keyword{internal}
