% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSOAregulart.R
\name{OSOAregulart}
\alias{OSOAregulart}
\title{TODO}
\usage{
OSOAregulart(s, k = 3, el = 3, m = NULL, permlist = NULL, random = TRUE)
}
\arguments{
\item{s}{TODO}

\item{k}{TODO}

\item{el}{TODO}

\item{m}{TODO}

\item{permlist}{TODO}

\item{random}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")

}
\keyword{internal}
