% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDLEs.R
\name{permopt}
\alias{permopt}
\title{optimize GWLP optimal oa for maximin criterion phi_p}
\usage{
permopt(oa, s, m)
}
\arguments{
\item{oa}{GWLP optimal orthogonal array}

\item{s}{TODO}

\item{m}{TODO}
}
\value{
a matrix of the same dimension as \code{oa}
}
\description{
optimize GWLP optimal oa for maximin criterion phi_p
}
\note{
not used any more
}
\keyword{internal}
