% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesGF.R
\name{poly2int}
\alias{poly2int}
\title{Galois field polynomial to an integer}
\usage{
poly2int(poly, gf)
}
\arguments{
\item{poly}{takes e x n matrix of polygon rows or a
length n vector for a single polygon (e=1)}

\item{gf}{Galois field object}
}
\value{
e element vector of integers
}
\description{
Galois field polynomial to an integer
}
\keyword{internal}
