% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.SOA}
\alias{print.SOA}
\alias{print.OSOA}
\alias{print.MDLE}
\title{Print Methods}
\usage{
\method{print}{SOA}(x, ...)

\method{print}{OSOA}(x, ...)

\method{print}{MDLE}(x, ...)
}
\arguments{
\item{x}{Strong Orthogonal Array}

\item{...}{Not used}
}
\description{
Print Methods
}
\examples{
## TODO
print(1)
}
