% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soa.R
\name{soa}
\alias{soa}
\title{TODO}
\usage{
soa(oa, t = 3, permlist = NULL, random = TRUE)
}
\arguments{
\item{oa}{TODO}

\item{t}{TODO}

\item{permlist}{TODO}

\item{random}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")

}
\keyword{internal}
