% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesCreate.R
\name{createSaturated}
\alias{createSaturated}
\alias{createAB}
\alias{BsFromB}
\alias{BcolsFromBcolllist}
\title{Utilities for array creation}
\usage{
createSaturated(s, k = 2)

createAB(s, k = 3, m = NULL)

BsFromB(B, s = NULL, r = NULL, permlist = NULL, oneonly = TRUE)

BcolsFromBcolllist(Bcollist)
}
\arguments{
\item{s}{levels}

\item{k}{integer; determines the run size: the resulting array will have s^k runs}

\item{m}{the number of columns to be created}

\item{B}{n x m matrix}

\item{r}{number of copies}

\item{permlist}{permutation list}

\item{oneonly}{logical: permute all copies in the same way?}

\item{Bcollist}{list of candidate columns for He et al. (2018) construction}
}
\value{
\code{createSaturated} returns an s^k times (s^k-1)/(s-1) matrix (saturated regular OA with s-level columns)

\code{createAB} returns a list of s^k times m matrices A, B and D

\code{BsFromB} returns an rn x m matrix

\code{BcolsFromBcolllist} returns column numbers selected for matrix B
}
\description{
Utilities for array creation
}
\examples{
createSaturated(3, k=3)  ## 27 x 13 array in 3 levels
}
\keyword{internal}
