% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShiTang.R
\name{create_ABC}
\alias{create_ABC}
\title{TODO}
\usage{
create_ABC(k, m = NULL, constr = c("ShiTang_alphabeta", "ShiTang_alpha"))
}
\arguments{
\item{k}{TODO}

\item{m}{TODO}

\item{constr}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")
}
