% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShiTang.R
\name{create_DfromABC}
\alias{create_DfromABC}
\title{TODO}
\usage{
create_DfromABC(listABC, permlist = NULL, random = FALSE, ...)
}
\arguments{
\item{listABC}{TODO}

\item{permlist}{TODO}

\item{random}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")
}
