% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDLEs.R
\name{DcFromDp}
\alias{DcFromDp}
\title{TODO}
\usage{
DcFromDp(
  Dp,
  s,
  ell,
  permlist = rep(list(rep(list(rep(0:(ell - 1), each = nrow(Dp)/(s * ell))), s)),
    ncol(Dp))
)
}
\arguments{
\item{Dp}{TODO}

\item{s}{TODO}

\item{ell}{TODO}

\item{permlist}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")
}
