% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nchoosek}
\alias{nchoosek}
\title{Utility functions from DoE.base}
\usage{
nchoosek(n, k)
}
\arguments{
\item{n}{number to select from}

\item{k}{number to be selected without replacement}
}
\value{
\code{nchoosek} returns a \code{k} times \code{choose(n,k)} matrix
whose columns hold the possible selections in lexicographic order
}
\description{
Utility functions from DoE.base
}
\keyword{internal}
