% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighbourcalcUniversal.R
\name{NeighbourcalcUniversal}
\alias{NeighbourcalcUniversal}
\title{TODO}
\usage{
NeighbourcalcUniversal(
  funname,
  mperm,
  r,
  ...,
  startperm = NULL,
  allpermlist = NULL,
  neighbordist = 1
)
}
\arguments{
\item{funname}{TODO}

\item{mperm}{TODO}

\item{r}{TODO}

\item{...}{TODO}

\item{startperm}{TODO}

\item{allpermlist}{TODO}

\item{neighbordist}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")

}
\keyword{internal}
