% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSOA_LiuLiut.R
\name{OSOA_LiuLiut}
\alias{OSOA_LiuLiut}
\title{Strong Orthogonal Arrays of Strength t using the method of Liu and Liu}
\usage{
OSOA_LiuLiut(oa, t = NULL, m = NULL, permlist = NULL, random = TRUE)
}
\arguments{
\item{oa}{input matrix or data.frame}

\item{t}{strength of \code{oa}.  If \code{NULL} (default), \code{t} will be
determined.  \code{t} can be chosen smaller than the input strength if
additional columns are desired.}

\item{m}{desired number of columns.  Defaults to the maximum possible.}

\item{permlist}{a list of length m of lists of length 1 of permutations of the levels}

\item{random}{a logical (a random draw of permutations is used if \code{permlist} is \code{NULL})}
}
\value{
an orthogonal array
}
\description{
Strong Orthogonal Arrays of Strength t using the method of Liu and Liu
}
\references{
Liu and Liu (2015)
}
\keyword{internal}
