% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSOAarbitrary.R
\name{OSOAarbitrary}
\alias{OSOAarbitrary}
\title{TODO}
\usage{
OSOAarbitrary(oa, el = 3, m = NULL, permlist = NULL, random = TRUE)
}
\arguments{
\item{oa}{TODO}

\item{el}{TODO}

\item{m}{TODO}

\item{permlist}{TODO}

\item{random}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")
}
