% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesEvaluate.R
\name{phi_p}
\alias{phi_p}
\title{Functions to evaluate uniformity of an array}
\usage{
phi_p(D, dmethod = "euclidean", p = 50, ...)
}
\arguments{
\item{D}{an array or an object of class SOA or MDLE}

\item{dmethod}{the distance to use, \code{"euclidean"} or \code{"manhattan"}}

\item{p}{the value for p to use in the formula for phi_p}

\item{...}{currently not used}
}
\value{
a number
}
\description{
phi_p calculates the discrepancy
}
\details{
small values of phi_p are associated with good performance on the
maximin distance criterion
}
\examples{
A <- DoE.base::L16.4.5  ## levels 1:4 for each factor
phi_p(A)
phi_p(A, method="manhattan")
A2 <- A
A2[,4] <- c(2,4,3,1)[A[,4]]
phi_p(A2, method="manhattan")
\dontrun{
  ## A2 has fewer minimal distances
  par(mfrow=c(2,1))
  hist(dist(A), xlim=c(2,6), ylim=c(0,40))
  hist(dist(A2), xlim=c(2,6), ylim=c(0,40))
}
}
\author{
Ulrike Groemping
}
