% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.SOA}
\alias{print.SOA}
\alias{print.OSOA}
\alias{print.MDLE}
\title{Print Methods}
\usage{
\method{print}{SOA}(x, ...)

\method{print}{OSOA}(x, ...)

\method{print}{MDLE}(x, ...)
}
\arguments{
\item{x}{object to be printed (SOA, OSOA, MDLE)}

\item{...}{further arguments for function \code{print}}
}
\description{
Print Methods
}
\examples{
myOSOA <- OSOAs_regular(s=3, k=3, optimize=FALSE)
print(myOSOA)
str(myOSOA)  ## structure for comparison
}
