% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.SOA}
\alias{print.SOA}
\alias{print.MDLE}
\alias{print.Spattern}
\alias{print.dim_wt_tab}
\title{Print Methods}
\usage{
\method{print}{SOA}(x, ...)

\method{print}{MDLE}(x, ...)

\method{print}{Spattern}(x, ...)

\method{print}{dim_wt_tab}(x, ...)
}
\arguments{
\item{x}{object to be printed (SOA, OSOA, MDLE, Spattern)}

\item{...}{further arguments for function \code{print}}
}
\value{
no value is returned
}
\description{
Print Methods
}
\examples{
myOSOA <- OSOAs_regular(s=3, k=3, optimize=FALSE)
myOSOA
str(myOSOA)  ## structure for comparison
Spat <- Spattern(myOSOA, s=3)
dim_wt_tab(Spat)
print(dim_wt_tab(Spat), na.print=" ")
}
