% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soa.R
\name{soa}
\alias{soa}
\title{work horse function for SOAs}
\usage{
soa(oa, t = 3, m = NULL, permlist = NULL, random = TRUE)
}
\arguments{
\item{oa}{TODO}

\item{t}{TODO}

\item{m}{TODO}

\item{permlist}{TODO}

\item{random}{TODO}
}
\value{
TODO
}
\description{
work horse function for SOAs
}
\examples{
print("TODO")

}
\keyword{internal}
