% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soa.R
\name{soa}
\alias{soa}
\title{work horse function for SOAs}
\usage{
soa(oa, t = 3, m = NULL, permlist = NULL, random = TRUE)
}
\arguments{
\item{oa}{matrix or data.frame that contains an ingoing symmetric OA. Levels must be denoted as 0 to s-1 or as 1 to s.}

\item{t}{the strength the SOA should have, can be 2, 3, 4, or 5. Must not
be larger than the strength of \code{oa}, but can be smaller. The resulting SOA will have s^t levels}

\item{m}{the requested number of columns (see details for permitted numbers of columns)}

\item{permlist}{list of lists of permutations}

\item{random}{logical}
}
\value{
a matrix
}
\description{
work horse function for SOAs
}
\details{
The function is the workhorse function for function \code{\link{SOAs}}.
}
\keyword{internal}
