% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiaoXuMDLE.R
\name{XiaoXuMDLE}
\alias{XiaoXuMDLE}
\title{TODO}
\usage{
XiaoXuMDLE(
  oa,
  ell,
  noptim.oa = 1,
  nseq = 2000,
  nrounds = 50,
  nsteps = 3000,
  dmethod = "manhattan",
  p = 50
)
}
\arguments{
\item{oa}{TODO}

\item{ell}{TODO}

\item{nseq}{TODO}

\item{nrounds}{TODO}

\item{nsteps}{TODO}

\item{optimize.oa}{TODO}
}
\value{
a matrix
}
\description{
TODO
}
\details{
deviates from Xiao and Xu by optimizing the ingoing OA for
phi_p instead of for the GWLP
}
\examples{
print("TODO")

}
\keyword{internal}
