% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiaoXuMDLE.R
\name{XiaoXuMDLE}
\alias{XiaoXuMDLE}
\title{TODO}
\usage{
XiaoXuMDLE(
  oa,
  ell,
  optimize.oa = TRUE,
  nseq = 2000,
  nrounds = 50,
  nsteps = 3000
)
}
\arguments{
\item{oa}{TODO}

\item{ell}{TODO}

\item{optimize.oa}{TODO}

\item{nseq}{TODO}

\item{nrounds}{TODO}

\item{nsteps}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
print("TODO")
}
